add latest reasonable suggestions to keep them, shutdown

+++ libraries/AP_Mount/AP_Mount_STorM32_native.h
@@ -61,7 +61,7 @@ private:
         TASK_SLOT4,
         TASK_SLOTNUMBER,
     };
-    uint64_t _task_time_last;
+    uint32_t _task_time_last;
     uint16_t _task_counter;
 
     // discovery functions
